requirements.txt
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
anthropic>=0.40.0
requests>=2.31.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
pytz>=2024.1
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.0
icalendar>=5.0.0
